refactor: extract AxisDrawer from GraphDrawer
